requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via fluent.runtime
babel==2.9.1
    # via fluent.runtime
fluent.runtime==0.3.1
    # via -r requirements.in
fluent.syntax==0.18.1
    # via fluent.runtime
pyicu==2.7.4
    # via -r requirements.in
pytz==2021.1
    # via
    #   babel
    #   fluent.runtime
six==1.16.0
    # via fluent.runtime
